+++ src/inspection/position/position.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+// Represents a cursor position in a text editor.
 export interface Position {
     readonly lineNumber: number;
     readonly lineCodeUnit: number;
